updated use cases.md
I updated use cases with use case numbers, triggering actions, and the flow of use cases is trackable by reading it. I also updated the use case diagram to show that the host is also a player of the game, since we decided to design the game so that we don't need a adult supervisor host.